# depot-lifecycle-openapi-2.2.5.yaml
openapi: 3.0.1
info:
  title: Depot Life Cycle
  description: "# Purpose\n\n A depot centric API for managing the interchange activity\
    \ & repair lifecycle of a shipping container.  The API is expected to be used\
    \ by Customers, Depots, and Owners to facilitate real time communication between\
    \ systems instead of traditional EDI files.\n\n\n\n # Overview\n\n The depot lifecycle\
    \ API is a RESTful API.  It's requests & responses are loosely based upon traditional\
    \ EDI files.  For code definitions, explanations, or traditional EDI definitions\
    \ refer to [IICL TB 002, February 2003](https://www.iicl.org/iiclforms/assets/File/public/bulletins/TB002_EDIS_February_2003.pdf).\n\
    \n\n\n # JSON\n\n The requests and responses of this API are formatted according\
    \ to the [JSON](https://www.json.org/) standard.\n\n ## Null vs Absent\n\n For\
    \ all types other than arrays, nullable and absent are equivalent by this specification.\
    \  For example, if a non-array attribute is marked optional it may be excluded\
    \ from the JSON or it may be explicitly set to null.  However, to keep request\
    \ sizes small, it is recommended to prefer excluding attributes rather than sending\
    \ an explicit null.\n\n For arrays, null should be avoided.  For example, if the\
    \ array is empty or there are no values expected for the array then either the\
    \ field can be excluded from the JSON or an empty array should be transmitted.\n\
...
  contact:
...
servers:
...
security:
- Dynamic_Token: []
- Static_Token: []
...
paths:
...
  /api/v2/estimate/{estimateNumber}:
    get:
      tags:
      - estimate
      summary: fetch an estimate revision
      description: "Finds an estimate by the given estimate number and depot, returning\
        \ the revision specified.  If revision is not specified, the current estimate\
        \ revision is returned."
      operationId: showEstimate
      parameters:
      - name: estimateNumber
        in: path
        description: the estimate number
        required: true
        schema:
          maxLength: 16
          type: string
          example: DEHAMCE1856373
      - name: depot
        in: query
        description: the identifier of the depot
        required: true
        schema:
          maxLength: 9
          pattern: "^[A-Z0-9]{9}$"
          type: string
          example: DEHAMCMRA
      - name: revision
        in: query
        description: the revision number to show for the estimate; when not specified
          the current revision will be returned.
        required: false
        schema:
          required:
          - "false"
          type: integer
          format: int32
          example: 0
      responses:
        "200":
          description: successfully found the estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Estimate"
        "400":
          description: invalid request was supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: fetching estimate is disallowed by security
        "404":
          description: estimate not found
        "501":
          description: this feature is not supported by this server
        "503":
          description: "API is temporarily paused, and not accepting any activity"
      security:
      - JWT_TOKEN: []
      x-iicl-purpose: reporting
    put:
      tags:
      - estimate
      summary: customer approve an estimate
      description: "Instead of sending in a full estimate revision, this endpoint\
        \ can be used to approve an estimate without revising it.  This endpoint would\
        \ typically be implemented by a lessee.  A depot would use the send revision\
        \ to report approvals."
      operationId: customerApproveEstimate
      parameters:
      - name: estimateNumber
        in: path
        description: the estimate number
        required: true
        schema:
          maxLength: 16
          type: string
          example: DEHAMCE1856373
      - name: depot
        in: query
        description: the identifier of the depot
        required: true
        schema:
          maxLength: 9
          pattern: "^[A-Z0-9]{9}$"
          type: string
          example: DEHAMCMRA
      requestBody:
        description: customer approval object to update an existing estimate
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimateCustomerApproval"
        required: true
      responses:
        "200":
          description: successfully approved the estimate revision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstimateAllocation"
        "201":
          description: successfully received the customer approval and issued a repair
            authorization for the estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
        "202":
          description: "customer approval accepted for processing, but not created\
            \ due to manual processing requirement"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingResponse"
        "400":
          description: an error occurred trying to customer approve the estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: customer approval is disallowed by security
        "404":
          description: the estimate or depot was not found
        "501":
          description: this feature is not supported by this server
        "503":
          description: "API is temporarily paused, and not accepting any activity"
      security:
      - JWT_TOKEN: []
      x-iicl-purpose: activity
    delete:
      tags:
      - estimate
      summary: cancels an estimate
      description: Cancels an estimate (deletes it and all revisions from the system)
      operationId: deleteEstimate
      parameters:
      - name: estimateNumber
        in: path
        description: the estimate number
        required: true
        schema:
          maxLength: 16
          type: string
          example: DEHAMCE1856373
      - name: depot
        in: query
        description: the identifier of the depot
        required: true
        schema:
          maxLength: 9
          pattern: "^[A-Z0-9]{9}$"
          type: string
          example: DEHAMCMRA
      responses:
        "200":
          description: successfully cancelled the estimate
          content:
            application/json:
              schema: {}
        "400":
          description: an error occurred trying to cancel the estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: cancel estimate is disallowed by security
        "404":
          description: the estimate or depot was not found
        "501":
          description: this feature is not supported by this server
        "503":
          description: "API is temporarily paused, and not accepting any activity"
      security:
      - JWT_TOKEN: []
      x-iicl-purpose: activity
    patch:
      tags:
      - estimate
      summary: update estimate totals
      description: "When the creation of the estimate is delayed due to a 202, after\
        \ the manual processing is complete, this method is called to perform the\
        \ update of the totals.  It is often implemented by a depot so that the lessor\
        \ can report updates."
      operationId: updateTotals
      parameters:
      - name: estimateNumber
        in: path
        description: the estimate number
        required: true
        schema:
          maxLength: 16
          type: string
          example: DEHAMCE1856373
      requestBody:
        description: total breakdowns to finish creating an estimate
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimateAllocation"
        required: true
      responses:
        "200":
          description: successfully received estimate totals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Estimate"
        "400":
          description: an error occurred trying to update totals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: estimate total update is disallowed by security
        "404":
          description: the estimate was not found
        "501":
          description: this feature is not supported by this server
        "503":
          description: "API is temporarily paused, and not accepting any activity"
      security:
      - JWT_TOKEN: []
      x-iicl-purpose: activity
  /api/v2/estimatePhoto/{relatedId}:
    post:
      tags:
      - estimate proposals
      summary: upload an estimate photo
      description: "Instead of using a link, upload a photo for a previously allocated\
        \ estimate"
      operationId: uploadEstimatePhoto
      parameters:
      - name: relatedId
        in: path
        description: the related identifier (from the estimate allocation) that this
          photo should be attached
        required: true
        schema:
          type: integer
          format: int64
          example: 10102561
      - name: line
        in: query
        description: an optional line number to associate this photo to
        schema:
          type: integer
          format: int32
          nullable: true
          example: 1
      - name: status
        in: query
        description: "indicator of when this photo applies\n\n`REPAIRED` - Photo is\
          \ after repair \n\n`BEFORE` - Photo is before repair"
        schema:
          maxLength: 8
          type: string
          nullable: true
          example: BEFORE
          default: BEFORE
          enum:
          - REPAIRED
          - BEFORE
      requestBody:
        description: The photo to upload (expected name of part is `file`)
        content:
          multipart/form-data:
            schema:
              type: string
              properties:
                file:
                  type: string
                  format: binary
              description: the photo data
              format: binary
            encoding:
              file:
                contentType: application/octet-stream
        required: true
      responses:
        "200":
          description: successfully uploaded an estimate photo
          content:
            application/json:
              schema: {}
        "202":
          description: "photo accepted for processing, but not created due to manual\
            \ processing requirement"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingResponse"
        "400":
          description: an invalid request was provided
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: uploading an estimate photo is disallowed by security
        "404":
          description: the estimate or line item could not be found
        "501":
          description: this feature is not supported by this server
        "503":
          description: "API is temporarily paused, and not accepting any activity"
      security:
      - JWT_TOKEN: []
      x-iicl-purpose: activity
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          maxLength: 6
          pattern: "^[A-Z0-9]{3}[0-9]{3}$"
          required:
          - "false"
          type: string
          description: indicator code specific to this error
          nullable: true
          example: TRI521
        message:
          required:
          - "false"
          type: string
          description: a descriptive error message
          nullable: true
          example: Info TRI521 - Unit has been gated-in but is not off-hired
        details:
          required:
          - "false"
          type: array
          description: details for the descriptive error message
          nullable: false
          items:
            type: string
            example: "Customer may turn in unit, but will continued to be billed."
      description: |-
        a general response object used when an error occurs to describe why it occurred

        `400` http status codes are used when a request is invalid or the basic requirements are not met

        `405` http status codes are used for business validations
    Estimate:
      required:
      - allocation
      - condition
      - currency
      - depot
      - estimateNumber
      - estimateTime
      - exchangeRate
      - total
      - unitNumber
      type: object
      properties:
        estimateNumber:
          maxLength: 16
          minLength: 1
          required:
          - "true"
          type: string
          description: The identifier for this estimate
          nullable: false
          example: DEHAMCE1856373
        unitNumber:
          maxLength: 11
          pattern: "^[A-Z]{4}[X0-9]{6}[A-Z0-9]{0,1}$"
          required:
          - "true"
          type: string
          description: The current unit number of the shipping container.
          nullable: false
          example: CONU1234561
        condition:
          description: |-
            an indicator on the status of the estimate and where it is in the revision process

            `D` - Damaged, Initial Estimate

            `E` - Customer Surveyed, No Approval

            `F` - Customer Approved Estimate, No Survey

            `G` - Customer Approved Estimate, Surveyed

            `L` - Owner Surveyed, No Approval
          allOf:
          - $ref: "#/components/schemas/EstimateCondition"
        estimateTime:
          required:
          - "true"
          type: string
          description: "the date and time of the estimate revision activity in local\
            \ time; i.e. `2020-07-21T17:32:28Z` \n\n( notation as defined by [RFC\
            \ 3339, section 5.6](https://tools.ietf.org/html/rfc3339#section-5.6)\
            \ )"
          format: date-time
          nullable: false
        comments:
          maxLength: 500
          required:
          - "false"
          type: string
          description: the comments concerning this estimate
          nullable: true
        requester:
          description: the party submitting this estimate
          nullable: true
          allOf:
          - $ref: "#/components/schemas/Party"
        depot:
          description: the location of this estimate
          allOf:
          - $ref: "#/components/schemas/Party"
        owner:
          description: the shipping container's owner
          nullable: true
          allOf:
          - $ref: "#/components/schemas/Party"
        customer:
          description: the lessee of the shipping container to whom customer line
            items should apply.
          nullable: true
          allOf:
          - $ref: "#/components/schemas/Party"
        currency:
          maxLength: 3
          pattern: "^[A-Z]{3}$"
          required:
          - "true"
          type: string
          description: the currency of amounts entered on this estimate
          nullable: false
          example: EUR
        total:
          required:
          - "true"
          type: number
          description: the sum of billed line items for this estimate
          format: double
          nullable: false
          example: 544.95
        exchangeRate:
          required:
          - "true"
          type: number
          description: the exchange rate to convert billed currency to the local currency
            of this estimate
          format: double
          nullable: false
          readOnly: true
          example: 0.8133
        customerApproval:
          description: lessee approval information for this estimate
          nullable: true
          allOf:
          - $ref: "#/components/schemas/EstimateCustomerApproval"
        type:
          description: |-
            delineates the type of estimate; whether it was the initial, primary estimate or an ancillary / secondary repair after the initial decision

            `R` - Primary (Unknown Estimation Standard)

            `RI` - Primary (IICL)

            `RC` - Primary (CWCA)

            `SC` - Secondary (CWCA)

            `SU` - Secondary Upgrade

            `AS` - Sell Upgrade

            `AU` - Ancillary Upgrade

            `AR` - Ancillary Repair
          nullable: true
          allOf:
          - $ref: "#/components/schemas/EstimateType"
          default: R
        upgradeType:
          description: |-
            the type of secondary upgrade this estimate represents.

            `FG` - Food grade

            `ML` - Malt

            `DB` - Dairy Board

            `EV` - Evian

            `WH` - Whiskey

            `SU` - Sugar

            `CF` - Coffee

            `TB` - Tobacco

            `MC` - Milk cartons

            `MP` - Milk powder

            `AM` - Ammunition

            `CH` - Cotton/Hay

            `TE` - Tea

            `FT` - Flexitank
          nullable: true
          example: AM
          allOf:
          - $ref: "#/components/schemas/UpgradeType"
        revision:
          required:
          - "false"
          type: integer
          description: the revision number of the estimate
          format: int32
          nullable: true
          example: 0
        lineItems:
          minItems: 0
          required:
          - "false"
          type: array
          description: detailed damage descriptions that when combined represent the
            damages being repaired by this estimate
          nullable: false
          items:
            $ref: "#/components/schemas/EstimateLineItem"
        photos:
          required:
          - "false"
          type: array
          description: An optional photo list for the shipping container damages.
          nullable: false
          items:
            $ref: "#/components/schemas/EstimatePhoto"
        allocation:
          description: the amount break downs by party for this estimate; typically
            only defined for a response and not included in requests
          readOnly: true
          allOf:
          - $ref: "#/components/schemas/EstimateAllocation"
      description: Represents an estimation of costs to repair or upgrade a shipping
        container.
    EstimateAllocation:
      required:
      - ctl
      - customerTotal
      - depot
      - estimateNumber
      - insuranceTotal
      - ownerTotal
      - revision
      - total
      type: object
      properties:
        relatedId:
          required:
          - "false"
          type: integer
          description: |-
            *Field is currently proposed to be added - not currently production approved.*

            An internal system identifier to be used to upload Estimate Photos or compare related activities.
          format: int64
          nullable: true
          example: 10102561
        estimateNumber:
          maxLength: 16
          minLength: 1
          required:
          - "true"
          type: string
          description: The estimate number this allocation applies to
          nullable: false
          example: DEHAMCE1856373
        revision:
          required:
          - "true"
          type: integer
          description: the revision number of the estimate this allocation applies
            to
          format: int32
          nullable: false
          example: 0
        depot:
          description: the location of the estimate
          allOf:
          - $ref: "#/components/schemas/Party"
        total:
          minimum: 0
          required:
          - "true"
          type: number
          description: the total for the estimate
          format: double
          nullable: false
          example: 544.95
        ownerTotal:
          minimum: 0
          required:
          - "true"
          type: number
          description: the total of the owner portion for the estimate
          format: double
          nullable: false
          example: 0
        customerTotal:
          minimum: 0
          required:
          - "true"
          type: number
          description: the total of the customer portion for the estimate
          format: double
          nullable: false
          example: 422.96
        insuranceTotal:
          minimum: 0
          required:
          - "true"
          type: number
          description: the total of the insurance portion for the estimate
          format: double
          nullable: false
          example: 121.99
        ctl:
          required:
          - "true"
          type: boolean
          description: indicates if the estimate causes the unit to be a constructive
            total lost
          nullable: false
          example: false
        comments:
          maxLength: 500
          required:
          - "false"
          type: string
          description: comments pertaining to the estimate creation
          nullable: true
          example: "Base Currency is: EUR; Based on estimate, user damages total:\
            \ 0.00; Calculated DV (base): 1151.84; Coverage amount: 121.99; Damage\
            \ exceeds DPP Coverage.; Damages exceed coverage amount. User pays excess.;\
            \ Total paid by DPP Coverage: 121.99; Total owed by user: 422.96;"
        preliminaryDecision:
          type: object
          description: "when possible, this is set to an expected sell/fix decision\
            \ to indicate the likely estimate owner approval action"
          nullable: true
          allOf:
          - $ref: "#/components/schemas/PreliminaryDecision"
          - type: object
      description: "Indicates the allocated estimate totals, including any comments\
        \ on the allocation and if the box is considered a constructive loss."
      readOnly: true
...
    EstimateCustomerApproval:
      required:
      - approvalDateTime
      - approvalTotal
      type: object
      properties:
        approvalNumber:
          maxLength: 128
          required:
          - "false"
          type: string
          description: The approval number from the customer
          nullable: true
          example: RAMON ROUBAL
        approvalDateTime:
          required:
          - "true"
          type: string
          description: "The time the estimate was approved by the customer; i.e. `2017-04-10T19:37:04Z`\
            \ \n\n( notation as defined by [RFC 3339, section 5.6](https://tools.ietf.org/html/rfc3339#section-5.6)\
            \ )"
          format: date-time
          nullable: false
        approvalUser:
          maxLength: 64
          required:
          - "false"
          type: string
          description: the user name or code who approved this estimate on behalf
            of the customer
          nullable: true
          example: KAERTS
        approvalTotal:
          minimum: 0
          required:
          - "true"
          type: number
          description: the amount in the currency of the estimate that's approved
            by the customer
          format: double
          nullable: false
          example: 422.96
      description: Represents the approval of the customer amount portion.
...
    PendingResponse:
      type: object
      properties:
        code:
          maxLength: 6
          pattern: "^[A-Z0-9]{3}[0-9]{3}$"
          required:
          - "false"
          type: string
          description: indicator code for this response
          nullable: true
          example: EXX365
        message:
          required:
          - "false"
          type: string
          description: a descriptive reason why the activity was not immediately created
          nullable: true
          example: "Info EXX365 - Accepted, but requires verification to process."
      description: used inform why an activity could not be immediately processed
...
    WorkOrder:
      required:
      - approvalDate
      - depot
      - lineItems
      - owner
      - type
      - workOrderNumber
      type: object
      properties:
        relatedId:
          required:
          - "false"
          type: integer
          description: |-
            *Field is currently proposed to be added - not currently production approved.*

            An internal system identifier to be used to upload Estimate Photos or compare related activities.
          format: int64
          nullable: true
          example: 10102561
        workOrderNumber:
          maxLength: 16
          minLength: 1
          required:
          - "true"
          type: string
          description: the identifier for this work order; this will be the approval
            number for repairs
          nullable: false
          example: WHAMG46019
        depot:
          description: the storage location where the shipping container is being
            repaired
          allOf:
          - $ref: "#/components/schemas/Party"
        owner:
          description: the owner of the shipping container
          allOf:
          - $ref: "#/components/schemas/Party"
        billingParty:
          description: the party that will bill the customer portion of damages for
            this repair
          nullable: true
          allOf:
          - $ref: "#/components/schemas/Party"
        type:
          maxLength: 11
          required:
          - "true"
          type: string
          description: the type of repair approved
          nullable: false
          example: SELLCWCA
        approvalDate:
          required:
          - "true"
          type: string
          description: |-
            the date and time in the depot local time zone (i.e. `2017-05-10T19:37:04Z`) that this work order is authorized for repair

            ( notation as defined by [RFC 3339, section 5.6](https://tools.ietf.org/html/rfc3339#section-5.6) )
          format: date-time
          nullable: false
        approvalTotal:
          required:
          - "false"
          type: number
          description: the total amount approved for repair
          format: double
          nullable: true
          example: 175
        approvalCurrency:
          maxLength: 3
          pattern: "^[A-Z]{3}$"
          required:
          - "false"
          type: string
          description: the currency of the approval total
          nullable: true
          example: EUR
        expirationDate:
          required:
          - "false"
          type: string
          description: |-
            the date and time in the depot local time zone (i.e. `2020-07-21T17:32:28Z`) that this repair must be completed by

            ( notation as defined by [RFC 3339, section 5.6](https://tools.ietf.org/html/rfc3339#section-5.6) )
          format: date-time
          nullable: true
        comments:
          maxLength: 500
          required:
          - "false"
          type: string
          description: comments pertaining to this repair for the intended recipient
            of this message
          nullable: true
          example: CWCA repairs for unit TCKU3456654 total 175.00 EUR per DEHAMCE1856373.1
        lineItems:
          maxItems: 200
          minItems: 1
          required:
          - "true"
          type: array
          description: units associated to this work order
          nullable: false
          items:
            $ref: "#/components/schemas/WorkOrderUnit"
      description: An approval of a damage estimate signifying a depot may repair
        a shipping container
...
  securitySchemes:
    JWT_TOKEN:
      type: http
      description: JWT Bearer Authentication
      scheme: bearer
      bearerFormat: JWT
...